@@ -1,6 +1,6 @@
 {
   "name": "koord2",
-  "version": "0.2.0",
+  "version": "0.4.0",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
@@ -9,54 +9,44 @@
     "lint": "vue-cli-service lint ./src"
   },
   "dependencies": {
-    "bootstrap-vue": "~2.22.0",
-    "chart.js": "^3.8.0",
-    "core-js": "^3.23.5",
-    "popper.js": "^1.16.1",
-    "portal-vue": "^2.1.7",
-    "uuid": "^8.3.2",
-    "vue": "^2.7.14",
-    "vue-class-component": "^7.2.6",
-    "vue-property-decorator": "^9.1.2",
-    "vue-router": "^3.6.5",
+    "chart.js": "^4.2.1",
+    "core-js": "^3.29.0",
+    "uuid": "^9.0.0",
+    "vue": "^3.2.47",
+    "vue-chartjs": "^5.2.0",
+    "vue-router": "^4.1.6",
     "vue-slider-component": "^3.2.24",
-    "vuetify": "^2.6.9",
-    "vuex-persist": "^3.1.3",
-    "vuex": "^3.6.2"
+    "vuetify": "^3.1.8",
+    "vuex": "^4.1.0",
+    "vuex-persist": "^3.1.3"
   },
   "devDependencies": {
     "@babel/polyfill": "^7.12.1",
     "@types/chart.js": "^2.9.37",
-    "@types/jest": "^27.5.2",
-    "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
-    "@typescript-eslint/typescript-estree": "^4.33.0",
-    "@vue/cli-plugin-babel": "^4.5.19",
-    "@vue/cli-plugin-eslint": "^4.5.19",
-    "@vue/cli-plugin-router": "^4.5.19",
-    "@vue/cli-plugin-typescript": "^4.5.19",
-    "@vue/cli-plugin-unit-jest": "^4.5.19",
-    "@vue/cli-plugin-vuex": "^4.5.19",
-    "@vue/cli-service": "^4.5.19",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/eslint-config-typescript": "^7.0.0",
-    "@vue/test-utils": "^1.3.0",
-    "babel-jest": "^27.5.1",
-    "bootstrap": "^4.6.2",
-    "chartjs-adapter-moment": "^1.0.0",
-    "eslint": "^6.8.0",
-    "eslint-plugin-prettier": "^3.4.1",
-    "eslint-plugin-vue": "^6.2.2",
+    "@types/jest": "^29.4.0",
+    "@types/uuid": "^9.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.54.1",
+    "@typescript-eslint/parser": "^5.54.1",
+    "@typescript-eslint/typescript-estree": "^5.54.1",
+    "@vue/cli-plugin-babel": "^5.0.8",
+    "@vue/cli-plugin-eslint": "^5.0.8",
+    "@vue/cli-plugin-router": "^5.0.8",
+    "@vue/cli-plugin-typescript": "^5.0.8",
+    "@vue/cli-plugin-unit-jest": "^5.0.8",
+    "@vue/cli-plugin-vuex": "^5.0.8",
+    "@vue/cli-service": "^5.0.8",
+    "@vue/eslint-config-prettier": "^7.1.0",
+    "@vue/eslint-config-typescript": "^11.0.2",
+    "@vue/test-utils": "^2.3.1",
+    "babel-jest": "^29.5.0",
+    "chartjs-adapter-moment": "^1.0.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-vue": "^9.9.0",
     "moment": "^2.29.4",
     "mutationobserver-shim": "^0.3.7",
-    "popper.js": "^1.16.1",
-    "portal-vue": "^2.1.7",
-    "prettier": "^2.7.1",
-    "sass": "^1.53.0",
-    "sass-loader": "^10.3.1",
-    "typescript": "~4.5.5",
-    "vue-cli-plugin-bootstrap-vue": "^0.8.2",
-    "vue-template-compiler": "^2.7.7"
+    "prettier": "^2.8.4",
+    "sass": "^1.58.3",
+    "sass-loader": "^13.2.0",
+    "typescript": "~4.9.5"
   }
 }
