@@ -1,7 +1,6 @@
 import "@babel/polyfill";
 import "mutationobserver-shim";
-import Vue from "vue";
-import "./plugins/bootstrap";
+import { createApp } from "vue";
 import "./plugins/chartjs";
 import vuetify from "./plugins/vuetify";
 import App from "./App.vue";
@@ -10,10 +9,9 @@ import store from "./store";
 
 import "@/assets/css/app.scss";
 
-// This is the main entry point of the application that initializes the Vue framework.
-new Vue({
-  router,
-  store,
-  vuetify,
-  render: (h) => h(App),
-}).$mount("#app");
+const app = createApp(App);
+
+app.use(router);
+app.use(store);
+app.use(vuetify);
+app.mount("#app");
