@@ -8,7 +8,7 @@ module.exports = {
     "eslint:recommended",
     "@vue/typescript/recommended",
     "@vue/prettier",
-    "@vue/prettier/@typescript-eslint",
+    '@vue/eslint-config-prettier'
   ],
   parserOptions: {
     ecmaVersion: 2020,
@@ -22,6 +22,9 @@ module.exports = {
         endOfLine: "auto",
       },
     ],
+    'vue/valid-v-slot': ['error', {
+      allowModifiers: true,
+    }],    
   },
   overrides: [
     {
