@@ -1,85 +1,79 @@
 <template>
-  <b-form @submit.prevent="authenticate">
-    <b-container fluid class="pl-0 mt-5">
-      <b-row>
-        <b-col cols="4">
-          <b-form-group
-            id="username"
+  <v-form @submit.prevent>
+    <v-container fluid class="pl-0 mt-5">
+      <v-row>
+        <v-col cols="4">
+          <v-text-field
+            v-model="username"
             label="User"
-            label-for="username-input"
-            :state="acceptUser"
-            invalid-feedback="Your username is too short."
-          >
-            <b-form-input
-              id="username-input"
-              v-model="username"
-              :disabled="disabled"
-              :state="acceptUser"
-              required
-            />
-          </b-form-group>
-        </b-col>
-      </b-row>
+            :disabled="disabled"
+            :rules="[acceptUser]"
+            required
+          />
+        </v-col>
+      </v-row>
 
-      <b-row>
-        <b-col cols="4">
-          <b-form-group
-            id="password"
+      <v-row>
+        <v-col cols="4">
+          <v-text-field
+            v-model="password"
             label="Password"
-            label-for="password-input"
-            :state="acceptPassword"
-            invalid-feedback="Your password is too short."
-          >
-            <b-form-input
-              id="password-input"
-              v-model="password"
-              type="password"
-              :disabled="disabled"
-              :state="acceptPassword"
-              required
-            />
-          </b-form-group>
-        </b-col>
-      </b-row>
-      <b-row class="mt-4">
-        <b-col>
-          <b-button
+            type="password"
+            :disabled="disabled"
+            :rules="[acceptPassword]"
+            required
+          />
+        </v-col>
+      </v-row>
+      <v-row class="mt-4">
+        <v-col>
+          <v-btn
             type="submit"
-            variant="primary"
+            color="primary"
+            @click="authenticate(username, password)"
             :disabled="disabled || !canAuthenticate"
-            >Log in</b-button
           >
-        </b-col>
-      </b-row>
-    </b-container>
-  </b-form>
+            Log in
+          </v-btn>
+        </v-col>
+      </v-row>
+    </v-container>
+  </v-form>
 </template>
 
 <script lang="ts">
-import { Vue, Component, Prop } from "vue-property-decorator";
+import { defineComponent } from "vue";
 
-@Component({})
-export default class LoginComponent extends Vue {
-  @Prop({ type: Boolean, default: () => false })
-  disabled!: boolean;
+export default defineComponent({
+  name: "LoginComponent",
+  props: {
+    disabled: {
+      type: Boolean,
+      default: () => false,
+    },
+  },
+  data() {
+    return {
+      username: "",
+      password: "",
+    };
+  },
+  computed: {
+    acceptPassword(): boolean {
+      return this.password.length > 0;
+    },
 
-  username = "";
-  password = "";
+    acceptUser(): boolean {
+      return this.username.length > 0;
+    },
 
-  get acceptPassword(): boolean {
-    return this.password.length > 0;
-  }
-
-  get acceptUser(): boolean {
-    return this.username.length > 0;
-  }
-
-  /**
-   * Returns true iff the login action should be enabled.
-   */
-  get canAuthenticate(): boolean {
-    return !this.disabled && this.username !== "" && this.password !== "";
-  }
+    /**
+     * Returns true iff the login action should be enabled.
+     */
+    canAuthenticate(): boolean {
+      return !this.disabled && this.username !== "" && this.password !== "";
+    },
+  },
 
   mounted(): void {
     const { VUE_APP_LIMESURVEY_LOGIN, VUE_APP_LIMESURVEY_PASSWORD } =
@@ -93,40 +87,42 @@ export default class LoginComponent extends Vue {
         );
       });
     }
-  }
+  },
 
-  private getCredentials(
-    login?: unknown,
-    password?: unknown
-  ): { password: string; username: string } {
-    const useLogin =
-      !login || typeof login !== "string" ? this.username : login;
-    const usePassword =
-      !password || typeof password !== "string" ? this.password : password;
-    return { username: useLogin, password: usePassword };
-  }
+  methods: {
+    getCredentials(
+      login?: unknown,
+      password?: unknown
+    ): { password: string; username: string } {
+      const useLogin =
+        !login || typeof login !== "string" ? this.username : login;
+      const usePassword =
+        !password || typeof password !== "string" ? this.password : password;
+      return { username: useLogin, password: usePassword };
+    },
 
-  /**
-   * Starts the authentication process with LimeSurvey.
-   *
-   * @param login the login to use
-   * @param password the password to use
-   */
-  async authenticate(login?: string, password?: string): Promise<void> {
-    this.$emit("auth-before", login);
-    try {
-      const credentials = this.getCredentials(login, password);
-      const okay = await this.$store.dispatch("authenticate", credentials);
-      if (okay) {
-        this.$emit("auth-success", login);
-      } else {
-        // noinspection ExceptionCaughtLocallyJS
-        throw new Error("Authentication failed");
+    /**
+     * Starts the authentication process with LimeSurvey.
+     *
+     * @param login the login to use
+     * @param password the password to use
+     */
+    async authenticate(login?: string, password?: string): Promise<void> {
+      this.$emit("auth-before", login);
+      try {
+        const credentials = this.getCredentials(login, password);
+        const okay = await this.$store.dispatch("authenticate", credentials);
+        if (okay) {
+          this.$emit("auth-success", login);
+        } else {
+          // noinspection ExceptionCaughtLocallyJS
+          throw new Error("Authentication failed");
+        }
+      } catch (e) {
+        this.$emit("auth-fail", login, e);
+        this.$store.commit("setError", e);
       }
-    } catch (e) {
-      this.$emit("auth-fail", login, e);
-      this.$store.commit("setError", e);
-    }
-  }
-}
+    },
+  },
+});
 </script>
