@@ -1,5 +1,14 @@
-// This is a general shim provided by Vue
+import VueRouter from "vue-router";
+
+/* eslint-disable */
 declare module "*.vue" {
-  import Vue from "vue";
-  export default Vue;
+  import type { DefineComponent } from "vue";
+  const component: DefineComponent<{}, {}, any>;
+  export default component;
+}
+
+declare module "vue/types/vue" {
+  interface Vue {
+    $router: VueRouter;
+  }
 }
