@@ -1,35 +1,35 @@
 <template>
-  <b-row class="home">
-    <b-col v-if="!isAuthenticated">
-      <b-container fluid class="pl-0">
-        <b-row>
-          <b-col cols="12">
+  <v-row class="home">
+    <v-col v-if="!isAuthenticated">
+      <v-container fluid class="pl-0">
+        <v-row>
+          <v-col cols="12">
             <h1>
               Not logged in yet
-              <b-icon icon="person" class="ml-4 text-primary"></b-icon>
+              <v-icon icon="person" class="ml-4 text-primary"></v-icon>
             </h1>
             <p class="lead mt-5">Welcome to the KoordTool.</p>
             <p>You can use this tool to visualize survey responses.</p>
 
             <p>Please authenticate first.</p>
-          </b-col>
+          </v-col>
 
-          <b-col cols="4">
-            <b-button to="/login" variant="outline-primary" class="mt-4">
+          <v-col cols="4">
+            <v-button to="/login" variant="outline-primary" class="mt-4">
               Authenticate
-            </b-button>
-          </b-col>
-        </b-row>
-      </b-container>
-    </b-col>
+            </v-button>
+          </v-col>
+        </v-row>
+      </v-container>
+    </v-col>
 
-    <b-col v-else>
-      <b-container fluid class="pl-0">
-        <b-row>
-          <b-col cols="12">
+    <v-col v-else>
+      <v-container fluid class="pl-0">
+        <v-row>
+          <v-col cols="12">
             <h1>
               Logged in as {{ username }}
-              <b-icon icon="person-check" class="ml-4 text-success"></b-icon>
+              <v-icon icon="person-check" class="ml-4 text-success"></v-icon>
             </h1>
             <p class="lead mt-5">Welcome to the KoordTool.</p>
             <p>You can use this tool to visualize survey responses.</p>
@@ -42,34 +42,35 @@
             >
               Limesurvey @ {{ instance }}
             </b-link>
-          </b-col>
-        </b-row>
+          </v-col>
+        </v-row>
         <survey-list :username="username" />
-      </b-container>
-    </b-col>
-  </b-row>
+      </v-container>
+    </v-col>
+  </v-row>
 </template>
 
 <script lang="ts">
-import { Vue, Component } from "vue-property-decorator";
 import SurveyList from "@/components/SurveyList.vue";
+import { defineComponent } from "vue";
 
-@Component({
+export default defineComponent({
+  name: "LoginView",
   components: {
     SurveyList,
   },
-})
-export default class HomeView extends Vue {
-  get isAuthenticated(): boolean {
-    return this.$store.getters.isAuthenticated;
-  }
+  computed: {
+    isAuthenticated(): boolean {
+      return this.$store.getters.isAuthenticated;
+    },
 
-  get username(): string {
-    return this.$store.getters.username;
-  }
+    username(): string {
+      return this.$store.getters.username;
+    },
 
-  get instance(): string {
-    return this.$store.getters.getInstanceDomain;
-  }
-}
+    instance(): string {
+      return this.$store.getters.getInstanceDomain;
+    },
+  },
+});
 </script>
