@@ -1,10 +1,10 @@
 <template>
   <div>
-    <b-row class="login">
-      <b-col> Currently connected to: {{ limeSurveyUri }} </b-col>
-    </b-row>
-    <b-row>
-      <b-col>
+    <v-row class="login">
+      <v-col> Currently connected to: {{ limeSurveyUri }} </v-col>
+    </v-row>
+    <v-row>
+      <v-col>
         <h2>Time slider interval in hours:</h2>
         <v-btn-toggle v-model="step" mandatory>
           <v-btn>
@@ -20,24 +20,28 @@
             <v-icon>24</v-icon>
           </v-btn>
         </v-btn-toggle>
-      </b-col>
-    </b-row>
+      </v-col>
+    </v-row>
   </div>
 </template>
 
 <script lang="ts">
-import { Vue, Component } from "vue-property-decorator";
+import { defineComponent } from "vue";
 
-@Component({})
-export default class SettingsView extends Vue {
-  get limeSurveyUri(): string {
-    return this.$store.getters.getLimeSurveyUri;
-  }
-  get step(): number {
-    return this.$store.getters.getStep;
-  }
-  set step(newStep: number) {
-    this.$store.commit("updateStep", newStep);
-  }
-}
+export default defineComponent({
+  name: "SettingsView",
+  computed: {
+    limeSurveyUri(): string {
+      return this.$store.getters.getLimeSurveyUri;
+    },
+    step: {
+      get(): number {
+        return this.$store.getters.getStep;
+      },
+      set(newStep: number) {
+        this.$store.commit("updateStep", newStep);
+      },
+    },
+  },
+});
 </script>
